half sqlite converted
