+++ packages/server/src/store/index.ts
@@ -24,9 +24,11 @@ export class Dataset extends Model<DatasetAttributes, DatasetCreationAttributes>
   public readonly updatedAt!: Date;
 }
 
+
 interface ProjectAttributes {
   id: string;
   name: string;
+  initialCheckpoint: string;
 
   epochs: number;
   batchSize: number;
@@ -40,11 +42,17 @@ type ProjectCreationAttributes = Optional<ProjectAttributes, keyof ProjectAttrib
 
 export class Project extends Model<ProjectAttributes, ProjectCreationAttributes> implements ProjectAttributes {
   name: string;
+  initialCheckpoint: string;
+
   epochs: number;
   batchSize: number;
   evalFrequency: number;
   percentEval: number;
+  
   inProgress: boolean;
+
+  //This only makes it compile, i dont know how to get all that ^^^^^^ into a hyperparamters "type".
+  hyperparameters: unknown;
 
   public readonly id!: string;
   public readonly createdAt!: Date;
@@ -84,6 +92,10 @@ Project.init(
     name: {
       type: new DataTypes.STRING(),
       allowNull: false
+    },
+    initialCheckpoint: {
+      type: new DataTypes.STRING(),
+      defaultValue: "default"
     },
     epochs: {
       type: DataTypes.INTEGER.UNSIGNED,
